+++ sharedPrompts/src/main/java/org/example/sharedprompts/global/config/scheduler/FallbackLockProvider.java
@@ -1,4 +1,4 @@
-package org.example.sharedprompts.global.config.shcduler;
+package org.example.sharedprompts.global.config.scheduler;
 
 import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.core.LockConfiguration;
